Add a fadein and fadeout segue methods.
Summary:
In the storyboard, create a custom type segue to destination view conroller, and set the identifier as "DOPFadeOutSegueForward",
It's all. run the appliation and tap on the photos, the timeline view will fadeout and segue forward to destination viewcontroller.
To segue back, in the destination ViewController, create a unwind connection with the Exit button.

+++ DOPTimelineView/DOPTimelineCommandsCell.m
@@ -116,7 +116,7 @@
 -(void) updateLikedByNum
 {
     if (_likedByNum > 0 ) {
-        self.likedNumLabel.text = [NSString stringWithFormat:@"%d",_likedByNum];
+        self.likedNumLabel.text = [NSString stringWithFormat:@"%ld",(long)_likedByNum];
         self.likedIcon.image = [UIImage likeIcon];
         
     }
@@ -125,7 +125,7 @@
 - (void) updateCommentedNum
 {
     if (_commentedNum > 0 ) {
-        self.commentedNumLabel.text = [NSString stringWithFormat:@"%d",_commentedNum];
+        self.commentedNumLabel.text = [NSString stringWithFormat:@"%ld",(long)_commentedNum];
         self.commentedIcon.image = [UIImage commentIcon];
         
     }
